Introduce notion of “dangerous” wasm unit tests
We have a unit test that is #if 0ed (and some additional unit tests incoming) that attempt to access out of bounds memory. WAVM handles this cleanly by trapping the SEGV and then we retire the transaction as a failure.

However in chain_test the Boost Test framework by default tries to grab SEGV for itself. So depending on the tests that execute WAVM will be unable to trap the SEGV and instead it will be bubbled up to Boost Test as a test failure.

I tried a few approaches to resolve this cleanly but long story short nothing great came about. There seems to be no way to programmatically disable the system fault handler without resorting to very complex modifications? What I’ve done is add a new test suite dangerous_wasm_tests that is by default disabled. When you run this test suite with WAVM you need to pass --catch_system_errors=no for the suite to pass deterministically. I’ve added the tests to ctest.